+++ lib/src/bloc_factories/base_bloc_default_factory.dart
@@ -1,0 +1,11 @@
+import 'package:custom_bloc/custom_bloc.dart';
+
+abstract class BaseBlocDefaultFactoryBloc<T, E> extends BaseBloc {
+  reload() {
+    initFetch();
+  }
+
+  initFetch() {}
+
+  initIfEmptyOrError() {}
+}