+++ lib/src/custom_stream_bloc.dart
@@ -16,7 +16,7 @@ typedef AsyncDataWidgetBuilder2<T, T2> = Widget Function(
     BuildContext context, T snapshot, T2 snapshot2);
 typedef AsyncDataWidgetBuilderMulti = Widget Function(
     BuildContext context, List<BaseModel> snapshots);
-typedef AsyncErrorWidgetBuilder<T, E> = Widget Function(
+typedef AsyncErrorWidgetBuilder<E> = Widget Function(
     BuildContext context, E error);
 typedef AsyncLoadingWidgetBuilder = Widget Function(BuildContext context);
 typedef AsyncNoContentWidgetBuilder = Widget Function(BuildContext context);
@@ -49,7 +49,7 @@ class CustomStreamBuilder<T, E, T2, E2> extends StatelessWidget {
   AsyncDataWidgetBuilderMulti? itemBuilderMulti;
 
   ///called when [addToError] is called on bloc
-  AsyncErrorWidgetBuilder<T, E>? errorBuilder;
+  AsyncErrorWidgetBuilder<E>? errorBuilder;
 
   ///called when [setAsLoading] is called on bloc
   AsyncLoadingWidgetBuilder? loadingBuilder;
@@ -201,85 +201,5 @@ class CustomStreamBuilder<T, E, T2, E2> extends StatelessWidget {
 
           return const SizedBox();
         });
-  }
-}
-
-typedef OnRetryCallback = Function();
-
-///A simple button that can be used to display errors
-class CustomBlocErrorWidget extends StatelessWidget {
-  final OnRetryCallback? onPressed;
-  final String buttonText;
-  final Color? buttonColor;
-  final String title;
-  final Widget? child;
-  final double buttonHeight;
-  final double maxButtonWidth;
-  final double buttonBorderRadius;
-  final Color buttonPressSplashColor;
-
-  const CustomBlocErrorWidget(
-      {Key? key,
-      this.onPressed,
-      this.buttonText = 'Retry',
-      this.buttonColor,
-      this.title = 'No item found',
-      this.child,
-      this.buttonHeight = 55,
-      this.maxButtonWidth = 450,
-      this.buttonBorderRadius = 10,
-      this.buttonPressSplashColor = Colors.black45})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    ThemeData theme = Theme.of(context);
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.center,
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        child ??
-            Column(crossAxisAlignment: CrossAxisAlignment.center, children: [
-              const SizedBox(
-                height: 34,
-              ),
-              Text(
-                title,
-                textAlign: TextAlign.center,
-                style: theme.textTheme.bodyMedium!
-                    .copyWith(fontWeight: FontWeight.w500),
-              ),
-              const SizedBox(
-                height: 14,
-              )
-            ]),
-        ConstrainedBox(
-          constraints: BoxConstraints(maxWidth: maxButtonWidth),
-          child: Material(
-            color: buttonColor ?? theme.buttonTheme.colorScheme!.primary,
-            borderRadius: BorderRadius.circular(buttonBorderRadius),
-            child: InkWell(
-              onTap: () {
-                if (onPressed != null) onPressed!();
-              },
-              borderRadius: BorderRadius.circular(buttonBorderRadius),
-              splashFactory: InkSplash.splashFactory,
-              splashColor: buttonPressSplashColor,
-              child: Container(
-                alignment: Alignment.center,
-                width: double.infinity,
-                height: buttonHeight,
-                child: Text(
-                  buttonText,
-                  style: theme.textTheme.labelLarge!
-                      .copyWith(fontWeight: FontWeight.w700, fontSize: 12),
-                ),
-              ),
-            ),
-          ),
-        ),
-      ],
-    );
   }
 }